% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Janie_functions.R
\name{read.wig}
\alias{read.wig}
\title{Read a .wig file}
\usage{
read.wig(data_file, start, end)
}
\arguments{
\item{data_file}{Path to the data file}

\item{start}{Start of the nucleotide range you would like to pull out}

\item{end}{End of the nucleotide range you would like to pull out}
}
\value{
A data frame with the nucleotide number and the reactivity
}
\description{
This function reads a .wig file into R
}
