% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Janie_functions.R
\name{stops.extract}
\alias{stops.extract}
\title{Quickly compiles stops data for analysis in other programs}
\usage{
stops.extract(
  file.wif,
  file.fasta,
  start,
  end,
  strand = "+",
  Name = "RNA",
  Nucleic_acid = "RNA",
  Winsorization = TRUE,
  Normalization_quantile = 0.95,
  Window = "center",
  Window_size = 10,
  Use_custom_windows = FALSE,
  custom_windows = list()
)
}
\arguments{
\item{file.wif}{Path to the .wif formatted file}

\item{file.fasta}{Path to the fasta file
#'@param start End of the nucleotide range you would like to pull out}

\item{end}{End of the nucleotide range you would like to pull out}

\item{strand}{Strand of the genome you want to analyze. Options: Default = "+" for the forward strand and "-" for the reverse strand}

\item{Name}{Name of the RNA. (This will be the prefix on the file that is saved). Default = "RNA".}

\item{Nucleic_acid}{The nucleic acid you want the output to be. Options: Default = "RNA" to print Us in the final file and "DNA" to print "Ts".}

\item{Winsorization}{Use the Winsorization method to normalize the data. Options: Default = TRUE to do the normalization or FALSE to skip the normalization.}

\item{Normalization_quantile}{Quartile you want to do the Winsorization normalization to.}

\item{Window}{Window type you want to use when calculating mean reactivities in sliding windows. Options: Default = "center" to apply the mean to the nucleotide at the center of the window, or "first" to apply the mean to the first nucleotide in the window.}

\item{Window_size}{Width of the sliding window that you want to apply. Default = 10 nucleotides}

\item{custum_windows}{A list of custum windows you want to calculate the mean for. Default = FALSE or a list of vectors in the format custom_windows = list(c(window1.start:window1.end), c(window2.start:window2.end)). Example: custom_windows = list(c(31:42), c(50:200)). Make sure that you use the relative location of the window on the RNA, not the absolute location on the genome.}
}
\value{
A csv file, a plot, and a dataframe
}
\description{
This function reads a .wig file and a fasta file into R, pulls out a specified subset of the data, cleans stop counts associated
with Gs and Us, Normalizes the data using the Winsorization method, calculates sliding window averages, and writes a csv file
for subsequent analysis. It also plots the data as a png for convenient assessments of the result.
}
